+++ src/pyg2p/__init__.py
@@ -1,6 +1,8 @@
-import logging
+version = (3, 0, 0)
+__authors__ = "Domenico Nappo"
+__version__ = 'v' + '.'.join(list(map(str, version)))
 
-__version__ = '3.0'
+import logging
 
 
 class Loggable: