+++ src/pyg2p/main/__init__.py
@@ -76,11 +76,6 @@ def config_command(conf, exc_ctx):
         conf.add_geopotential(geo)
         logger.info(f'Added geopotential {geo} to configuration')
 
-    elif exc_ctx.run_tests:  # -t
-        # comparison tests (grib2pcraster vs pyg2p, pyg2p scipy interpol vs pyg2p GRIBAPI interpol)
-        from pyg2p.main.testrunner import TestRunner
-        TestRunner(conf.tests.vars, exc_ctx.get('test.cmds')).run()
-
     elif exc_ctx.check_conf:  # -K
         # check unused intertables (intertables that are not in configuration and can be deleted
         conf.check_conf()