+++ tests/__init__.py
@@ -11,7 +11,7 @@ def check_dataset(self, ds):
     result_dir = self.options['results'].joinpath(f'{ds}')
     reference_dir = self.options['reference'].joinpath(f'{ds}')
     comparator = PCRComparator()
-    diffs = comparator.compare_dirs(reference_dir.as_posix(), result_dir.as_posix())
+    diffs = comparator.compare_dirs(reference_dir.as_posix(), result_dir.as_posix(), skip_missing=False)
     if diffs:
         logger.info(diffs)
     assert not diffs